+++ src/deepquantum/photonic/__init__.py
@@ -16,7 +16,7 @@ from . import state
 from . import torontonian_
 
 from .ansatz import Clements, GaussianBosonSampling, GBS_Graph
-from .circuit import QumodeCircuit, QumodeCircuitTDM
+from .circuit import QumodeCircuit
 from .decompose import UnitaryDecomposer
 from .draw import DrawClements
 from .gate import PhaseShift, BeamSplitter, MZI, BeamSplitterTheta, BeamSplitterPhi, BeamSplitterSingle, UAnyGate
@@ -26,6 +26,7 @@ from .mapper import UnitaryMapper
 from .measurement import Generaldyne, Homodyne
 from .qmath import permanent, takagi, xxpp_to_xpxp, xpxp_to_xxpp, quadrature_to_ladder, ladder_to_quadrature
 from .state import FockState, GaussianState
+from .tdm import QumodeCircuitTDM
 from .torontonian_ import torontonian
 from .utils import set_hbar, set_kappa
