+++ tests/test_photonic_fock.py
@@ -68,12 +68,12 @@ def test_batched_fock_basis_states_and_data():
         res3 = cir(data=data[i], state=init_state[i], is_prob=True)
 
         # test is_prob=None
-        assert torch.equal(res1, re1[i])
+        assert torch.allclose(res1, re1[i], atol=1e-6)
 
         for key in res2.keys():
             # test is_prob=False
-            assert torch.allclose(res2[key], re2[key][i], atol=1e-6)
+            assert torch.allclose(res2[key], re2[key][i], atol=1e-5)
 
         for key in res3.keys():
-            # test is prob = True
+            # test is prob=True
             assert torch.allclose(res3[key], re3[key][i], atol=1e-6)
